# Parallel configure/activate in lifecycle manager