--- a/nav2_lifecycle_manager/include/nav2_lifecycle_manager/lifecycle_manager.hpp
+++ b/nav2_lifecycle_manager/include/nav2_lifecycle_manager/lifecycle_manager.hpp
@@ -259,6 +259,7 @@ protected:
   // Whether to automatically start up the system
   bool autostart_;
   bool attempt_respawn_reconnection_;
+  bool parallel_state_transitions_;
 
   NodeState managed_nodes_state_{NodeState::UNCONFIGURED};
   diagnostic_updater::Updater diagnostics_updater_;