@jplapp 2025-10-17
So most of the "slow launch" impact was caused by having used a bond_heartbeat_period of 1.0. There is some remaining speedup but I'm not sure if it's worth the additional complexity. Let me know, then I'll update the docs if needed.