@SteveMacenski 2025-10-21
I want to make sure I'm understanding you correctly: you don't think this is worthwhile anymore due to finding the issue with your bringup speed was something else? If so, then I think we can close