@jplapp 2025-10-21
Yes, basically activating+configuring stage goes from 3 seconds to 2 seconds. Personally I wouldn't add the additional parameter and logic for that. Thanks for your review and quick feedback! Hope that next time I can propose some more meaningful speedups :smile: